Todo-master/backend/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2023.7.22
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
pep8==1.7.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
sqlparse==0.4.4
toml==0.10.2
urllib3==1.26.16
